Split UserListItemVM into dedicated file
