+++ OTS.Solution/OTS.DOMAIN/MobileAccountingVM/UserListItemVM.cs
@@ -1,0 +1,14 @@
+namespace OTS.DOMAIN.MobileAccountingVM
+{
+    public class UserListItemVM
+    {
+        public string? UserName { get; set; }
+        public string? Email { get; set; }
+        public int UserId { get; set; }
+        public bool EmailConfirmed { get; set; }
+        public string? Role { get; set; }
+        public int? Head { get; set; }
+        public string? Id { get; set; }
+        public string? Status { get; set; }
+    }
+}
