+++ OTS.Solution/OTS.DOMAIN/MobileAccountingVM/UserRequestVM.cs
@@ -78,15 +78,4 @@ namespace OTS.DOMAIN.MobileAccountingVM
         public bool IsActive { get; set; }
     }
 
-    public class UserListItemVM
-    {
-        public string? UserName { get; set; }
-        public string? Email { get; set; }
-        public int UserId { get; set; }
-        public bool EmailConfirmed { get; set; }
-        public string? Role { get; set; }
-        public int? Head { get; set; }
-        public string? Id { get; set; }
-        public string? Status { get; set; }
-    }
 }
